requirements.txt
python-telegram-bot==21.4
google-api-python-client==2.154.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
python-dotenv==1.0.1
SQLAlchemy==2.0.23
matplotlib==3.9.2
seaborn==0.13.2
